Fix C4244(Truncation loss of data) and C4245(Signed/Unsigned mismatch) warnings of MSVC

+++ include/sys/types.h
@@ -22,7 +22,7 @@ typedef unsigned short blksize_t;
 typedef unsigned int blkcnt_t;
 typedef unsigned long long fsblkcnt_t;
 typedef unsigned long long fsfilcnt_t;
-typedef long long suseconds_t;
+typedef long suseconds_t;
 typedef unsigned long long useconds_t;
 
 #endif
